+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
   config property `requestedSecretLifetime`. This helps reduce frequent Pod restarts ([#796]).
 - Run a `containerdebug` process in the background of each Kafka container to collect debugging information ([#803]).
 - Aggregate emitted Kubernetes events on the CustomResources ([#809]).
+- Support configuring JVM arguments ([#819]).
 
 ### Changed
 
@@ -25,6 +26,7 @@ All notable changes to this project will be documented in this file.
 [#809]: https://github.com/stackabletech/kafka-operator/pull/809
 [#810]: https://github.com/stackabletech/kafka-operator/pull/810
 [#813]: https://github.com/stackabletech/kafka-operator/pull/813
+[#819]: https://github.com/stackabletech/kafka-operator/pull/819
 
 ## [24.11.1] - 2025-01-10